requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
mysql-connector-python==8.3.0

# Vector Store & ML
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Document Processing
pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0

# LangChain & AI
langchain==0.1.6
langgraph==0.0.20
openai==1.10.0

# Utilities
python-dotenv==1.0.1
numpy==1.26.3